Remove compatibility code from BaseCoordinateFrame
The signature of `BaseCoordinateFrame.represent_as()` was changed in a
backwards-incompatible way in b688c0f1da1e469d188311729c2eb15ea1793cae
(change argument order of represent_as and add deprecation warning), but
compatibility code was added to allow calls following the older
signature to succeed with a warning. This compatibility no longer serves
a purpose because the new signature is old enough by now.