Update `pre-commit` configuration
The update is done manually for two reasons. Ruff tried to remove usage
of deprecated `numpy` functions, but the usage was in tests that were
checking that those function calls work. Furthermore, `codespell` had
some trouble figuring out how to fix a couple of typos and one of the
newly detected typos is actually a name. [ci skip]