Fix bug for integer compressed images
In _tiled_compression.py:decompress_image_data_section(), blanked
pixels are assigned the value np.nan. That is incorrect for integer
datatypes and raises an exception because np.nan cannot be converted
to an integer.

The fix is to use the BLANK keyword from the header as a fill value
rather than np.nan. ZBLANK is used in place of BLANK if no BLANK
keyword is found. Note the the FITS standard recommends (but does
not require) that BLANK should be used instead of ZBLANK for integer
images. The new code works correctly with either ZBLANK or BLANK
in the compressed image header.

Several modules have been changed to implement the correct behavior.
The changes only apply to integer images.

compressed.py:CompImageHDU.__init__()

Save the BLANK value from the header (or ZBLANK if BLANK is not
specified) to the self._blank attribute for integer images.

If necessary, a BLANK keyword with value of the most negative integer
(e.g., -32768 for 16-bit integers) is added to the header. That will
only happen in the rare case that neither BLANK nor ZBLANK is specified
in the header, but the compressed image table includes a ZBLANK column.
The header ought to have a header keyword with BLANK for that case
(otherwise why is ZBLANK there at all?), but it is not clear from the
FITS standard that a missing BLANK keyword should be treated as an
error. An AstropyUserWarning is issued in that case.

compressed.py:CompImageHDU._scale_data()

Apply the conversion to float if self._blank is defined, even if
BSCALE/BZERO are omitted. That allows the array that is returned to the
user to use np.nan to indicate blank pixels. This is consistent with the
behavior of uncompressed FITS integer images.

Note this conversion to float applies to unsigned integer images, which
differs from the uncompressed image behavior. In discussion we agreed
that the different behavior for ints and uints is incorrect and should
be fixed for uncompressed images. (Previously compressed uint images
with blanks simply did not work at all.)

This code also fixes a bug where the image datatype was changed (to
float or uint, depending on the parameters) before comparing to the
BLANK value to find blanks. That is potentially incorrect for float
types (e.g., the BLANK value could be a 64-bit integer that is not
exactly represented as a float64). And it is definitely incorrect for
the unsigned types, where the blank value is supposed to be tested
before the conversion to unsigned. For example, a typical value for
BLANK for an int16 image is -32768, and the comparison to that negative
value does not work correctly if the image has already been converted to
uint16. In the new code, the tests for blanks is done first, and then
the image is converted to the new datatype. Note this also allowed
removing the mysterious conversion of the blank value to int32.

The np.where() statement converts uint datatypes to float so that NaN can
be used for blanks. Coerce the np.nan value to float32 so that float32
is used for the output array if possible. Note that np.where will still
create a float64 array if the uint size is larger (uint32 or uint64).
This ensures that the datatype is the same for the uint and int
datatypes when they are converted to a float datatype. This version
works for all datatypes.

header.py:_bintable_header_to_image_header()

If an image has ZBLANK in the header but not BLANK, ensure that the
uncompressed header has the BLANK keyword so that blanks are properly
handled. Since using ZBLANK instead of BLANK is permitted by the FITS
standard, this is not treated as an error and no warning is issued.

_tiled_compression.py:decompress_image_data_section()

The zblank_header value can come from either the ZBLANK or BLANK
keyword. For integer images, the zblank_header value is used to fill
blanks instead of np.nan. Note that the change described above in
compressed.py:CompImageHDU.__init__() guarantees that either
ZBLANK or BLANK is present in the header.

tests/test_tiled_compression.py:test_integer_blank_support()

This new function tests blank support for integer images. A test integer
image with blanks is created from the m13.fits image. Modified versions
of that compressed file are also created by editing the binary table
extension to use ZBLANK instead of BLANK and to use a ZBLANK column in
the table instead of a header value. All the files are read and the
results checked to confirm correct pixel values. The presence of a BLANK
keyword in the uncompressed header is also confirmed (regardless of how
the BLANK value was specified in the compressed image). And the expected
warning if a BLANK value had to be invented is captured and confirmed.

docs/changes/io.fits/16550.bugfix.rst

Change log