Add more variety to integer blank tests
Change the value used for the BLANK keyword to -16384 instead of using
the default value of -32768. That could catch errors in the default
assignment.

Add an additional test using an image with a ZBLANK column plus a BLANK
keyword in the header. That is a standard configuration and is processed
without any warnings.