Simplify `codespell` configuration about names
There are two files that list contributors to `astropy`. Some of the
names might look like misspelled words, so `codespell` should be
configured to ignore them. Previously this was done by ignoring such
names, but that could lead to genuine typos going undetected elsewhere
and might require maintainer effort in case new names get added or if
the `codespell` word list gets updated. It is better to configure
`codespell` to ignore the two files instead. This does allow typos in
them to go unnoticed, but the amount of text is very small.

Setting up per-line `codespell` ignores is not suitable because the
BibTeX entries are meant to be copy-pasteable. [ci skip]