Improve many unit formatter parametrized test ids
Test ids are displayed when `pytest` is called with the `--verbose`
option or for failed tests. In some cases assigning good ids was
simplified by implementing a simple `NamedTuple` for storing test data,
and implementing a helper function for generating those `NamedTuple`
instances aviods needless code churn that would otherwise occurr because
of changing the indentation levels of parametrizations.