Update test_projections.py
I just walked by and saw your tests. I think I could help you a little. Here are my suggestions:
- Refactor out code into helper functions and reuse them.
- At the same time make the tests shorter and easier to read.
- Don't use files.
- Split the tests into fast unit tests and slow functional tests. Like this, the tests will be run much more often and are more helpful.

I didn't get to run the tests so far on my PC so I don't know the exact contents of the header. This has to be fixed along with the name of the setup function and removing comments.

Let me know if you appreciate my help or if I should continue walking.