Speed up `OGIP` unit parser for composite units
`OGIP._functions` stores the functions that the OGIP standard allows.
Previously it was a `list`, but now it is a `frozenset`, so simply
checking if an element is in it is now faster. Furthermore, the parser
now avoids needlessly checking for functions.

A new unit tests ensures a new code branch is tested.