Ensure power of a unit is not a `numpy` number
In several places `astropy.units` checks if a power of a unit is `int`,
`float` or `fractions.Fraction` instance. This means that a
`numpy.floating` instance takes a different code path than a `float` and
a `numpy.integer` takes a different code path than an `int`.
`astropy.units.utils.sanitize_power()` was meant to prevent such
confusing behaviour, but it missed several edge cases. Those edge cases
are now fixed and the new `UnitPower` type alias helps with type
checking.