Implement `UnitBase.components()`
The new method provides a convenient way of looping over the bases and
powers of a unit, which previously required accessing both its `bases`
and `powers` properties. In some cases this avoids having to store the
unit in a variable so that the two properties could be called
separately. It also enables specifying `zip()` with `strict=True` once
in the definition of `components` instead of having to write out
`strict` every time `bases` and `powers` are zipped together.