Make the `CDS` unit formatter a `Generic` subclass
`CDS` and `Generic` both implement string parsing, so making `CDS` a
subclass of `Generic` can help avoid code duplication. More refactoring
is needed to fully take advantage of this, but the `_get_unit()`
implementation in `CDS` can be already be safely deleted. One
complication is that `CDS` does need to inherit its
`format_exponential_notation()` method from `Base`, not `Generic`, but
the simplifications will make up for it.