Make `NamedUnit.get_format_name()` private
The method is error-prone because it returns a value even for incorrect
`format` arguments. Furthermore, `CompositeUnit` does not have the
method, so it can only be used with some units. The `to_string` method
has neither of these problems, so that is what users should be calling
instead. The public `get_format_name()` should be replaced with a
private method and the public version should be deprecated.