Split `sanitize_scale()` unit utility
The scale of a `CompositeUnit` should be sanitized. Some checks have to
be done after computations involving the scale are complete, but others
can be done before computations. The previous `sanitize_scale()` utility
could perform redundant checks at runtime if it was called more than
once, but the new `sanitize_scale_type()` and `sanitize_scale_value()`
functions allow performing the checks only at the appropriate times.
Furthermore, checking that scale is not zero is now done by
`sanitize_scale_type()`, not by `CompositeUnit.__init__()`.

The changes did cause one existing test to fail, but it was safe to
remove because the behaviour it checks is not promised by our API.