Make parsing units with `Unit()` thread-safe
Parsing units from strings with `Unit()` can involve
`warnings.catch_warnings()`, which is not thread-safe. This could cause
problems because the machinery can be invoked when reading files and
speeding up I/O with threading is common enough. The solution is to use
a `threading.RLock`. `RLock` is preferable over `Lock` because e.g. the
`Generic` parser calls `Unit()`, so there might be more than one attempt
per thread to obtain the lock.