Solve problems with Ruff formatting in a file
Ruff is somewhat picky about the placement of `fmt: off`, `fmt: on` and
`fmt: skip` instructions, which can cause formatting to be accidentally
turned off. The Ruff rule RUF028 would guard against this, but it is
still in preview. The command `ruff check --select RUF028 --preview`
revealed problems in one file.