Detect negative powers without parenthesis in OGIP
The OGIP unit parser now detects negative powers without parenthesis,
which are not allowed by the OGIP standard. Changing the `SIGN` token in
the OGIP lexer allows the error message about missing parenthesis to
refer to the power exactly as provided by the user, otherwise integer
powers would be written as floats in the message.