Turn CDS, OGIP and VOUnit into FITS subclasses
Currently all four listed formatters are subclasses of `Generic`, which
implements several methods for its subclasses that it does not use
itself. Making `FITS` the common parent class of `CDS`, `OGIP` and
`VOUnit` will enable decluttering `Generic` without introducing needless
code duplication in its subclasses. The only minor complication is that
the `Base` formatter needs to move the implementation of `to_string()`
to `_to_string()` so that `FITS.to_string()` could be bypassed.