Make u.Unit(..., format="vounit") obey parse_strict
The VOUnit standard allows users to define custom units, so the
`astropy` `VOUnit` parser emitted a warning when it encountered a
non-standard unit instead of raising an error. However, this behaviour
could not be modified with the `parse_strict` parameter of `u.Unit()`,
which has made it needlessly complicated for users to validate their
unit strings and at the same time it has also made it complicated to
suppress the warnings. Changing the type of the emitted warning invokes
the machinery introduced in c3831694dfecc3ad61a2b478aa2a025c0ebfdb66
(Implement mechanism for more flexible unit parsing) and addresses both
problems.

This does not affect custom units that have been explicitly declared
by quoting them.