MAINT: factor out parser methods to a new mixin class.
This way all classes that provide parsers can rely on the same set of
standard methods to be available.  Note that little attempt was made
to actually reduce their number.

In the process, also refactor generic, fits, vounit, which all need
the same parser, but have different treatment of strings around it.
So, easiest to just define the parser in a new class, which they can
inherit from.