Simplify and speed up `_UnitMetaClass.__call__()`
Most importantly the case of `represents` being a `Quantity` does not
have to be handled by `_UnitMetaClass` explicitly because it is handled
by `Unit.__init__()` anyways. Furthermore, if `represents` is not `None`
then `s` will be handed to `NamedUnit.__init__()`, which does not accept
`Quantity` instances. Fewer code branches are needed in `_UnitMetaClass`
and the common case of `str` input can be handled earlier, which speeds
it up.

A small change in the definition of `percent` avoids an import loop.