Define unit formatter protocols
Previously the `Base` unit formatter not only implemented functionality
common to the unit formatters but also defined the public API. This
necessitated the presence of a skeleton `Base.parse()`, which
complicated refactoring and made the logic in `_known_formats()`
fragile, but could not be changed without sacrificing public API
stability. The simplest solution (since Python 3.8) is to define the
public API in protocols so that the methods in `Base` provide one
possible implementation of the API rather than its definition.