Fix customizing the cache directory
657c91a53e7bf049a7e29e9910bcbdd8453bd3dd (MNT: Run PTH flake test in
prep for supporting pathlib (config)) introduced a bug which caused the
`XDG_CACHE_HOME` environment variable to be ignored and
`XDG_CONFIG_HOME` to control both the configuration and the cache
directories (if set). The correct behaviour has been restored, but
removing any cache files from unintended locations is not attempted.