Remove three useless properties from `UnitBase`
The purpose of the `names`, `name` and `aliases` properties of
`UnitBase` is to produce more helpful error messages if a user tries to
access them on units that are not `NamedUnit` instances. In practice the
presence of these properties makes them appear in autocompletion, which
is misleading for users, and they also confuse type checkers.