Deprecate arithmetic with units and str or bytes
Automatically converting `str` or `bytes` instances to units when they
are involved in products or multiplications with a unit is fragile
because products involving numerical values, strings and units can
produce, depending on the order of the multiplicands or the placement of
parenthesis, an `UnitBase`, a `Quantity` or an error, even though
multiplication is supposed to be associative and commutative. The
`isinstance(..., (bytes, str))` checks slow down code that makes no use
of this automatic conversion. The functionality is also unnecessary
because using units directly or converting strings to units explicitly
is simple enough.