Read LaTeX tables where there is whitespace before the `\begin{tabular}` (#17624)
* Read LaTeX tables where there is whitespace before the `\begin{tabular}`

This commit  fixes #4595 as discussed in that issue:
It makes the finding of the \begin{tabular} (and similar keywords) a little bit more flexible by allowing extra white space. Extra white space is easy to introduce when copying and pasting from a LaTeX file. It does not implement any further flexibility. It does not allow extra text in front of the tag because the reader cannot replace a full LaTeX parser an we prefer to fail in cases like `1 & 3 & 4 \\ \end{tabular}` instead of silently dropping the last row of data just because it happens to be in the same line as the `\end{tabular}`.
It also does not add a reader to read LaTeX tables without the \bgin{tabular} \end{tabular} lines. Every full LaTeX table has those and if users copy and paste from a longer LaTeX file, they can simply include those lines. Documentation has been added to that effect.