Split the `test_si_prefixes()` unit test
The previous test checked the prefix names and symbols together, but it
is better to check them separately because the number of recognized
names and symbols of a prefix can be different. For example, the names
"deca" and "deka" are recognized as aliases for the same prefix, which
nonetheless has a single symbol "da".