Make unit prefixes immutable
The variables containing SI and binary prefixes are conceptually
constants, so they should use immutable data structures. Defining the
`UnitPrefix` class as a subclass of `NamedTuple` achieves immutability
and also provides user convenience because of named attributes.

The new class is public so that it could be referred to in public
annotations and docstrings, but because `UnitPrefix` refers to
`UnitScale` then the latter must be made public too.