Allow importing Å, ℓ and Ω units
The `Generic` parser already knew that Å, ℓ and Ω mean angström, litre
and ohm respectively, but it was not possible to import those units
using those symbols from the `units` namespace. Now the Unicode symbols
are proper aliases for the units, which also means that the `Generic`
parser does not have to special-case them.