Add regression test for a unit name collision bug
When a unit is injected to a namespace, `astropy` does check for name
collisions. However, the namespace often belongs to a module and then
the expectation is that the unit can be accessed as an attribute of the
module with the name being the identifier, in which case Python applies
the NFKC normalization to the name. The new test reveals that the
`astropy` collision detection does not account for the normalization.