Check unit name collisions with NFKC normalization
`astropy` has machinery for injecting units to a namespace. The
namespace often belongs to a module and then the expectation is that the
unit can be accessed as an attribute of the module with the name being
the identifier, in which case Python applies the NFKC normalization to
the name. The injection machinery now applies the normalization to unit
names when checking for name collisions.