Stop storing unit scale as `Fraction` or `int`
The scale factor of a unit is now a `float` if it is real or a `complex`
otherwise. Previously it was also allowed to be a `Fraction` or an
`int`. Those types are still allowed for creating a unit, but they will
be converted to a `float`. In practice they were converted to a `float`
in many cases already.

Limiting the types of a scale factor makes it simpler to write code to
handle them. Most notably, testing with `np.testing.assert_allclose()`
works just fine with `float`, but not `Fraction` or very large `int`.