Enable units from `required_by_vounits` properly
In pull request #5661 it was decided that the prefixed versions of
`solLum`, `solMass` and `solRad` units should not be present in the main
`units` namespace. However, they are valid units in the VOUnit format,
so they are defined in a separate `required_by_vounit` module.
Furthermore, because the `"generic"` format is supposed to be a superset
of the other formats then parsing those units should work with the
`"generic"` format too. Previously the `required_by_vounit` module added
its units to the registry the `"generic"` format uses, but that was a
side-effect of `required_by_vounit` being imported, which happened by
using the `"vounit"` format. Now the `"generic"` formatter can be used
to parse those units even if `"vounit"` has not been used yet.