TST: Re-enable test_all for devdeps but only when noscipy is not specified. We have to copy over some stuff from test_all in pyproject.toml to tox.ini because tox unable to grab them automatically when there are multiple extras.
Because of the tox bug, test_all does nothing for
recdeps and alldeps in extras in tox.ini so we remove
the misleading directive. In fact, if you install the rest of the
stuff in test_all into alldeps via the deps directive,
oldestdeps start to fail because it silently upgrades to latest
numpy instead of using the minversion of numpy.