Rewrite `required_by_vounit` tests
The single test for `required_by_vounit` is replaced with multiple
parametrized tests, which are better because now a failing `assert` does
not prevent running the other checks. Equally importantly the tests are
moved out from `test_deprecated.py` because nothing in
`required_by_vounit` is nor should be deprecated.