Update invalid unit formatter error messages
If the user specifies an invalid format when converting strings to units
or units to strings then it is helpful if the error message lists the
valid names. However, some formatters can be only be used for formatting
units, while others can also be used for parsing strings. Previously the
code created the same error message in both cases, which contained
separate lists with the parsing and formatting-only formatter names. It
is more user-friendly if the parsing error only lists the parsing
formatters and the formatting error lists the formatters in a single
list.