Remove `PhysicalType.__len__()`
Previously a `PhysicalType` instance exposed the number of names it has
in its `__len__()`, but this was not documented anywhere and the meaning
of `len(length)` being `1` (by default) is not obvious. It is not
possible to deprecate the method because that could lead to deprecation
warnings that the user cannot avoid (e.g. if they call `list(length)`).