Define degrees Rankine as scaled version of kelvin
The Kelvin, Celsius and Fahrenheit temperature scales all have different
zero-points, which means the normal unit conversion machinery cannot
handle them. But the Rankine scale has the same zero-point as the Kelvin
scale, so redefining degrees Rankine as a scaled version of the kelvin
removes the need for special-casing it like degrees Celsius and degrees
Fahrenheit.