Simplify imports in unit definition modules
At some time in the past the unit definition modules did not define
`__all__` explicitly, which meant that anything imported had to be
imported using an alias that started with an underscore, imported inside
a function to stay out of module-level namespace, or cleaned up
explicitly. Such tricks are no longer needed.