Ensure SI unit names start with a lowercase letter
SI unit names start with a lowercase letter even if their symbol begins
with a capital letter. `astropy.units.si` has followed that guideline
for some units, but for others it has defined names with both lowercase
and uppercase first letters. The names with uppercase letters have now
been removed with three exceptions: `Angstrom` is not an SI unit at all
and should perhaps be moved to a different file entirely, `Celsius`
should really be named `degrees_Celsius`, and `Ohm`, which might be
justified because it is often used as an ASCII-compatible symbol for the
unit.