Deprecate hashing scalar `Quantity` instances
Hashing scalar `Quantity` instances is fundamentally broken because the
hashes of equal quantities might nonetheless be different. As a special
case of that, the hash of a `Quantity` can change during its lifetime,
which allows it to be inserted more than once into a `set` or a `dict`.
It is best to deprecate the functionality so that it could be removed in
the future. Hashing non-scalar quantities is already impossible.