Adding a utf8trunc function to io.votable.util.
This would be necessary once we prototype the interpretation of char
as utf-8 sequence.

Apologies for just dumping in a bunch of doctests and possibly ignoring
all kinds of style guides; if nobody else volunteers to clean this up, I'll
take another dive into the astropy tooling once I've pulled my development
machine to trixie.