Explain why HTML readers might not be good choice for journal tables
In #14346 it was noted that the HTML table reader does not easily return a table from an MNRAS article. Deeper study shows that there are problems with any major journal, due to inconsistent mark-ups, captchas, changing formats, embedded JS, or that tables are only included as jpegs in some cases. Thus, it seems impossible to make a working table parser for even the top few journals (see details of discussion in  #14346).
This PR tries ot summarize the findings in the API and narrative docs in the locations that I hope users are most likely to see it. The discussion of the original issue suggested more extreme measures (such as rasing a warning every time the HTML reader is used or deprecating it altogether) but I deem that too invasive, since the reader itself works and is useful for a number of legitimate cases, including the case where a user manually downloads the HTML for an article and reads it after verifying that the table in question is actually contained in the file.

closes #14346

[skip CI]