Avoid stringified type annotations in `units`
Sphinx is known to have problems resolving stringified annotations, so
it's best to avoid them whenever possible. In `units` it is not possible
to avoid stringified annotations entirely because of circular imports,
but it was nonetheless possible to remove `if TYPE_CHECKING:` blocks
from all but two modules. Moving the definitions of `UnitBase` and
`PhysicalType` was necessary for minimizing the number of forward
references.