Improve duplicate unit definition error message
The machinery that injects units into a namespace checks for name
collisions. Because Python applies NFKC normalization to identifiers
then a name collision can occur even if before normalization the names
are different. The new error message is less cluttered if normalization
is not required for collision and more informative if it is.