ENH: Allow setup.py to create unit stubs on installation
This moves the previous script to a new file in typing_utils
and adjusts it so that it can be called from setup.py on
installation (in regular or editable mode). Test are adjusted
accordingly.

Since typing is at present not directly useful, and IDEs will
find these stubs automatically, no announcement of this is
made in the changelog or documentation.