Refactor unit formatter `_validate_unit()` methods
The `"vounit"` and `"ogip"` unit formats have deprecated units. The code
for handling deprecated units in those two formatters is now replaced
with a common implementation in `_ParsingFormatMixin`.

The `"generic"` format does not have to worry about deprecated units, so
it does not benefit from the code that is being added to
`_ParsingFormatMixin`. In fact calling `super()._validate_unit(...)` in
`Generic._validate_unit()` would slow it down (slightly).