Improve validating unit module __all__ entries
The unit definition modules define so many objects that maintaining the
`__all__` for them manually is not feasible. Instead the vast majority
of entries are generated by `generate_dunder_all()`. The function now
checks that everything in its output is a valid identifier. At least in
`cds` this previously was not the case.