Allow handling warnings when writing deprecated units
The `"ogip"` and `"vounit"` formats have deprecated units. Converting
such units to strings has so far caused an unavoidable warning. A new
parameter of the `UnitBase.to_string()` and unit formatter `to_string()`
methods allows silencing that warning or raising an error instead. This
is analogous to the `parse_strict` parameter of unit parsing methods.

The `StrEnum` that controls the new behavior is defined in a new module
so that it could be imported without having to worry about import loops.