Allow automatically converting deprecated units
The "vounit" and "ogip" formats have deprecated units. It is now
possible to instruct the unit formatters to try to replace deprecated
units when writing them to strings. If automatic replacement is not
possible the deprecated unit is used and a warning is emitted.

An alternative would have been to implement a new method in `UnitBase`
to replace deprecated units, which could be used like
`u.erg.convert_deprecated("vounit").to_string(format="vounit")`
but then the user would have to specify the unit format twice and such
repetition should be avoided.