Define a Greenwich `EarthLocation` for tests
Several tests require an `EarthLocation` instance for testing some other
functionality. Currently many of them create an `EarthLocation`
representing Greenwich using the private `get_builtin_sites()` function,
but it is much simpler to define a private `EarthLocation` instance.
This completely removes the need to use `get_builtin_sites()` in tests
outside `astropy/coordinates/tests/test_sites.py`. That file is
different from the rest because it is meant to test the site registry,
so it should not be using any pre-defined `EarthLocation` instances.