Avoid `get_builtin_sites()` in all tests
The `get_builtin_sites()` function was used (both directly and via
`EarthLocation._get_site_registry(force_builtin=True)`) in some of the
tests to enable testing the `EarthLocation` site registry without having
to worry about fetching data from the Internet, but in practice it meant
that the code that was executed in the tests was different from the code
users would execute at runtime. Now the tests avoid
`get_builtin_sites()` entirely. Two of the rewritten tests `xfail`
because they catch the bug reported in
https://github.com/astropy/astropy/issues/18572