Remove the builtin `EarthLocation` site registry
`EarthLocation` can maintain a registry of known observatory sites. The
data for that registry is not bundled with `astropy`. So far, if
`astropy` could not find cached data and could not fetch the data from
the Internet then as a last resort `astropy` fell back on a minimal
registry that contained Greenwich as the single entry. Given that such a
tiny registry is not likely to be useful for users, it is not worth
keeping.