Fix simple-to-address typing errors in coordinates
The changes here halve the number of errors Mypy reports if invoked with
`mypy --follow-imports silent --exclude tests/ -- astropy/coordinates/`
In most cases the cause of the reported error was a missing annotation,
which caused Mypy to automatically infer a type that was too narrow.
Sometimes the correct place for the annotation was a parent class
outside `coordinates`.