Improving the new kwarg `loc` feature (rebased squash of multiple commits)
Reverting to default verticalalignment value set to

Updating user warning's stacklevel to 2

Revising `loc` support for `'c'` axis
Future updates should reconsider the purpose of `loc` in the case of an `EllipticalFrame` after the label drawing is fixed for axis `'c'`

Refining the documentation of `set_axislabel` to detail the dictionary configurability of its visual property parameters

Replacing use of try/except statements in visual property getters

Changing the default of the `loc` kwarg to `None`
This will set it to the corresponding entry in the `pyplot.rcParams` dictionary - which for a clean installation should be `'center'`

Updating the signature to not include `loc`
The `loc`` kwarg does not need to be explicitely included in the signature, and explicitely set via `set_loc``, since the `set` method inherited from `Artist` will if kwargs contains the `loc` entry. Explicitely having `loc` included and set via `set_loc` results in an unintended behavior that will overwrite `loc` if calling multiple time `set_axislabel` without specifying `loc`.

Fixing the `axis` and `loc` conditional statement in `draw`
Updating the NotImplementedError raising case for unsupported `loc` values, from flagging if `loc` is not `center`, to flagging if `loc` is not None