Remove the `_dev` sub-package
`astropy` uses `setuptools_scm` to generate version information
dynamically instead of hardcoding it in `pyproject.toml`. However,
`setuptools_scm` generates the version data when `astropy` is built, so
in the case of an editable install (like a developer is likely to use)
the version data can still become outdated. The `_dev` sub-package was
supposed to address that problem, but it requires `setuptools_scm` to be
installed and that is usually not the case because `setuptools_scm` is a
build-time, not a run-time dependency. If `setuptools_scm` happens to be
installed then it can be invoked directly with the command
```
python -m setuptools_scm
```
so even in that case `_dev` is not helpful.