Generate stubs for CDS and Imperial units
The `cds` and `imperial` modules differ from the other unit definition
modules because they also define `enable()` functions. The stub
generation machinery can now write stubs for functions without
parameters, which is good enough for the modules in question. The return
types of the functions are looked up from the annotations of the
implementations, which avoids needless hard-coding in the stub
generation machinery.